requirements.txt
aiogram==3.13.1
paramiko==3.5.0
psycopg2-binary==2.9.10

